- renamed backup_of_trunk_before_switch_to_codecompletion_refactrong (/branches/backup_of_trunk_before_switch_to_codecompletion_refactrong ) to backup_of_trunk_before_switch_to_codecompletion_refactoring.

+++ src/sdk/wxscintilla/src/scintilla/lexers/LexMagik.cxx
@@ -17,6 +17,7 @@
 #include "Scintilla.h"
 #include "SciLexer.h"
 
+#include "PropSetSimple.h"
 #include "WordList.h"
 #include "LexAccessor.h"
 #include "Accessor.h"
@@ -233,7 +234,7 @@ static void ColouriseMagikDoc(unsigned int startPos, int length, int initStyle,
 	            }
 
 	            if(characters.InList(keyword)) {
-	                sc.Forward(static_cast<int>(strlen(keyword)));
+	                sc.Forward(strlen(keyword));
 	            } else {
 	                sc.Forward();
 	            }
